Add template commit message file to .gitconfig
Unfortunately the path of the template file had to be hard coded
because .gitconfig does not support variable expansion.

If I found a way, I would use the XDG_CONFIG_HOME variable, but
at least I got to use the tilde expansion.

This should not be a huge problem because if the XDG_CONFIG_HOME gets
changed and the 'git/message' file moves to another place, 'git commit'
will notify the user about it.