Make dotfiles structure simpler to handle and more intuitive
* Merge apps/ and cfgs/

The '.dotfiles' extension was added to prevent potential conflicts
inside the .config/ folder when applications decide to use this
folder too.

* Convert '.dotfiles' karabiner cfg to native cfg

* Rename 'iterm' to 'iterm2'

Because that's how they are called in Homebrew.

* Rename 'sublime' to 'sublime-text'

Because that's how they are called in Homebrew.

* Convert iterm2's app setup to a regular setup

* Convert sublime's app setup to a regular setup

* Group home-based dotfiles in the 'home' folder

* Rename 'bins' to 'bin'

To make the dotfiles' structure more intuitive.

* Rename 'cfgs' to 'config'

To make the dotfiles' structure more intuitive.

* Remove custom app setups from 'setup-apps.sh'

They are now implemented via regular setups.

* Update the bin's setup

* Update config's setup

* Update home's setup

* Update the main setup script