Update lualine separators in Neovim
The default lualine separators looked cropped because of the way
Alacritty renders Nerd Font icons with non-zero 'font.offset' and
'font.glyph_offset', so we switched to simpler separators which look
great in Alacritty.

It's worth noting that in order to get best looking separators, you need
to enable 'font.builtin_box_drawing' in Alacritty which causes the
vertical line drawing character to stretch.